+++ lib/model/hourly_forecast.dart
@@ -9,7 +9,7 @@ class HourlyForecast {
   final int humidity;
   final double dewPoint;
   final int cloudiness;
-  final int uvIndex;
+  final double uvIndex;
   final double precipitationProb;
   final WindData windData;
   final WeatherCondition weatherCondition;
